### 2025-01-09 #168 @editorialbot
[PRE REVIEW]: Extrae.jl: 
**Submitting author:** @mofeing (<a href="http://orcid.org/0000-0003-4273-3328">Sergio Sánchez Ramírez</a>)
**Repository:** https://github.com/bsc-quantic/Extrae.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
**Managing EiC:** Luca Ferranti

Thanks for submitting your paper to JCON @mofeing. **Currently, there isn't a JCON editor assigned** to your paper.
