### 2025-01-19 #168 @mofeing
@luraess one question: changes like this should just be made to the repo where i have the paper uploaded (in this case, https://github.com/bsc-quantic/Extrae.jl) or do i need to submit it again?
