### 2025-01-19 #168 @luraess
> [@luraess](https://github.com/luraess) one question: changes like this should just be made to the repo where i have the paper uploaded (in this case, https://github.com/bsc-quantic/Extrae.jl) or do i need to submit it again?

No need to submit again. Ideally you'd provide a fix and ask the editorialbot to generate again the PDF. Thanks!
