### 2025-01-20 89bced3
Format bibliography file
