+++ paper/ref.bib
@@ -53,35 +53,35 @@
 }
 
 @inproceedings{10.1145/3404397.3404440,
-author = {Munera, Adrian and Royuela, Sara and Llort, Germ\'{a}n and Mercadal, Estanislao and Wartel, Franck and Qui\~{n}ones, Eduardo},
-title = {Experiences on the characterization of parallel applications in embedded systems with Extrae/Paraver},
-year = {2020},
-isbn = {9781450388160},
-publisher = {Association for Computing Machinery},
-address = {New York, NY, USA},
-url = {https://doi.org/10.1145/3404397.3404440},
-doi = {10.1145/3404397.3404440},
-abstract = {Cutting-edge functionalities in embedded systems require the use of parallel architectures to meet their performance requirements. This imposes the introduction of a new layer in the software stacks of embedded systems: the parallel programming model. Unfortunately, the tools used to analyze embedded systems fall short to characterize the performance of parallel applications at a parallel programming model level, and correlate this with information about non-functional requirements such as real-time, energy, memory usage, etc. HPC tools, like Extrae, are designed with that level of abstraction in mind, but their main focus is on performance evaluation. Overall, providing insightful information about the performance of parallel embedded applications at the parallel programming model level, and relate it to the non-functional requirements, is of paramount importance to fully exploit the performance capabilities of parallel embedded architectures. This paper contributes to the state-of-the-art of analysis tools for embedded systems by: (1) analyzing the particular constraints of embedded systems compared to HPC systems (e.g., static setting, restricted memory, limited drivers) to support HPC analysis tools; (2) porting Extrae, a powerful tracing tool from the HPC domain, to the GR740 platform, a SoC used in the space domain; and (3) augmenting Extrae with new features needed to correlate the parallel execution with the following non-functional requirements: energy, temperature and memory usage. Finally, the paper presents the usefulness of Extrae to characterize OpenMP applications and its non-functional requirements, evaluating different aspects of the applications running in the GR740.},
-booktitle = {Proceedings of the 49th International Conference on Parallel Processing},
-articleno = {53},
-numpages = {11},
-keywords = {performance evaluation, parallel programming models, analysis tools, OpenMP, Embedded systems},
-location = {Edmonton, AB, Canada},
-series = {ICPP '20}
+  author = {Munera, Adrian and Royuela, Sara and Llort, Germ\'{a}n and Mercadal, Estanislao and Wartel, Franck and Qui\~{n}ones, Eduardo},
+  title = {Experiences on the characterization of parallel applications in embedded systems with Extrae/Paraver},
+  year = {2020},
+  isbn = {9781450388160},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  url = {https://doi.org/10.1145/3404397.3404440},
+  doi = {10.1145/3404397.3404440},
[…]
