+++ paper/ref.bib
@@ -6,7 +6,8 @@
   number={1},
   pages={65--98},
   year={2017},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1137/141000671}
 }
 
 @inproceedings{pillet1995paraver,
@@ -27,7 +28,8 @@
   number={1},
   pages={3--60},
   year={2011},
-  publisher={SAGE Publications Sage UK: London, England}
+  publisher={SAGE Publications Sage UK: London, England},
+  doi={10.1177/1094342010391989}
 }
 
 @article{wagner2017performance,
@@ -37,7 +39,8 @@
   volume={108},
   pages={2171--2179},
   year={2017},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.procs.2017.05.203}
 }
 
 @inproceedings{1592857,
@@ -77,7 +80,7 @@
   pages = {107472},
   year = {2025},
   issn = {0167-739X},
-  doi = {https://doi.org/10.1016/j.future.2024.07.050},
+  doi = {10.1016/j.future.2024.07.050},
   url = {https://www.sciencedirect.com/science/article/pii/S0167739X24004187},
   author = {Brian J.N. Wylie and Judit Giménez and Christian Feld and Markus Geimer and Germán Llort and Sandra Mendez and Estanislao Mercadal and Anke Visser and Marta García-Gasulla},
   keywords = {Hybrid parallel programming, MPI message-passing, OpenMP multithreading, OpenACC device offload acceleration, HPC application execution performance measurement & analysis, Performance assessment & optimisation methodology & tools, Hands-on training & coaching},
@@ -222,7 +225,8 @@
   volume={3},
   pages={32--36},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.softx.2015.10.004}
 }
 
 @article{browne2000portable,
@@ -233,5 +237,6 @@
   number={3},
   pages={189--204},
   year={2000},
-  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA},
+  doi={10.1177/109434200001400303}
 }
