### 2025-01-20 #168 @luraess
@mofeing you may need to remove the square brackets in the yml file for author affiliation to fix the error.
