### 2025-01-20 #168 @luraess
Yes - thanks! And PDF generation may be fixed upon formatting correctly the yaml file.
