### 2025-01-20 5b30b9f
Fix affiliation syntax in paper
