### 2025-01-20 #168 @luraess
Seems we're hitting
```
Package caption Warning: Unknown document class (or package),
(caption)                standard defaults will be used.
See the caption package documentation for explanation.

)

! LaTeX Error: Command \subcaption already defined.
               Or name \end... illegal, see p.192 of the manual.
```
in https://github.com/JuliaCon/proceedings-papers/actions/runs/12874309577/job/35893496045#step:3:222
