### 2025-01-20 #168 @mofeing
yes, the problem is that the JuliaCon document class defines `\subcaption` here
https://github.com/JuliaCon/JuliaConSubmission.jl/blob/8f2d1e7fc169d9622606b418768f429ce71bb390/paper/juliacon.cls#L138-L142

and it clashes with the `subcaption` package that i use for the subfigures...

let's see if undefining it before loading `subcaption` works
