### 2025-01-20 #168 @mofeing
the problem now is that i'm using `svg` and i need to manually include the generated PDFs
