### 2025-01-21 85bd0ce
Try avoiding svg package and manually include generated TeX code
