+++ paper/paper.tex
@@ -10,7 +10,7 @@
 \usepackage{soul}
 \usepackage{amssymb}
 \usepackage{graphicx}
-\usepackage{svg}
+% \usepackage{svg}
 
 \newcommand{\jlpkg}[1]{\texttt{#1}}
 \newcommand{\extraejl}{\jlpkg{Extrae.jl}}
@@ -306,7 +306,8 @@ Furthermore, the variability is small enough compared to the amount of time to d
 
 \begin{figure}[h]
     \centering
-    \includesvg[width=\linewidth]{mpi_call_profile.svg}
+    % \includesvg[width=\linewidth]{mpi_call_profile.svg}
+    \input{mpi_call_profile_svg-tex.pdf_tex}
     \caption{Fraction of time spent on each MPI routine. Dispersion comes from the time spent by different MPI ranks.}
     \label{fig:mpi-runtime-perc}
 \end{figure}
