### 2025-01-21 9d4d848
Fix width of SVG figure
