### 2025-01-21 7aa300a
Fix undefined `\svgwidth` length
