+++ paper/paper.tex
@@ -307,6 +307,7 @@ Furthermore, the variability is small enough compared to the amount of time to d
 \begin{figure}[h]
     \centering
     % \includesvg[width=\linewidth]{mpi_call_profile.svg}
+    \newlength{\svgwidth}
     \setlength{\svgwidth}{\linewidth}
     \input{mpi_call_profile_svg-tex.pdf_tex}
     \let\svgwidth\undefined
