### 2025-02-01 #168 @luraess
Thanks both @JBlaschke and @carstenbauer for accepting to review this manuscript!

@JBlaschke the way it works; the next step there will be a GH issue opened being the revision and you'll generate a check-list there for your to-dos. Ideally, you give a succinct feedback in the review issue and open PRs in the package of interest for details, issues, and suggested modifications. You can find most of the details in the [Proceedings Guide](https://juliacon.github.io/proceedings-guide/reviewer/).
