### 2025-02-01 #180 @editorialbot
[REVIEW]: Extrae.jl: 
**Submitting author:** @mofeing (<a href="http://orcid.org/0000-0003-4273-3328">Sergio Sánchez Ramírez</a>)
**Repository:** https://github.com/bsc-quantic/Extrae.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
