### 2025-02-05 #180 @carstenbauer
Hi all, while I haven't completed all of my review works (for example, I haven't actually tested Extrae.jl yet), here comes my first batch of comments:

#### CONFLICT OF INTEREST
* @luraess, For clarity reasons, I should point out that I'm mentioned in the acknowledgements of the paper. However, I have not collaborated with the authors on Extrae.jl beyond talking/brainstorming about it.

#### GENERAL CHECKS
* Regarding "Contribution and authorship": Looking at the repo contributions, it's not obvious why @clasqui (Marc Clascà) isn't an author but @giordano (Mosè Giordano) is. Can you shortly comment on this?

#### DOCUMENTATION
* Statement of need is missing
* No installation instructions. Not for Extrae.jl and not for the parent tool (at least a link?).
* "Example usage": Would be great to have a few more examples (e.g. multithreading, MPI).
* "Automated tests": No CI setup.
* "Community guidelines": Not available.

#### PAPER
* See comments/notes in the **attached PDF file**: [10.21105.jcon.00180.pdf](https://github.com/user-attachments/files/18669392/10.21105.jcon.00180.pdf)
* Expand statement of need: Why do we need Extrae, specifically? Why aren't the other, mentioned tools not sufficient?

#### OTHER COMMENTS
* Link to JuliaCon24 talk in the README.md?
