### 2025-02-11 #180 @mofeing
> Regarding "Contribution and authorship": Looking at the repo contributions, it's not obvious why @clasqui (Marc Clascà) isn't an author but @giordano (Mosè Giordano) is. Can you shortly comment on this?

Sure. The development of Extrae.jl was carried out on different stages. I first started Extrae.jl to learn about Julia and BinaryBuilder, but I didn't really use it much.

A while later, @clasqui contacted me as he saw my package and was interested in using it for profiling an app that used Distributed.jl. He added a custom `ClusterManager` that did the same thing as the SSH Cluster Manager but emitted Extrae events. I think that worked for him and we stopped developing it more.

Later, I did a major refactor of the code to keep it simpler and to try to make it compatible with new developments in Tracepoints.jl (currently known ProbeBase.jl) and DistributedNext.jl (which adds a hooking system which would let us do what @clasqui was looking for). This included removing the custom `ClusterManager`.

So the code added by @clasqui, excluding bugfixes, didn't make it to the final version presented at JuliaCon but I'm happy to add him as co-author if reviewers see it appropiate.

The development of Extrae.jl was resumed when @giordano came as a visitor to my research group. We were both looking to make Extrae work on A64FX to profile code on Fujitsu machines. His work was mostly performed on the background, as he got in charge of fixing the Yggdrasil build for Extrae_jll. It was a time-consuming task because:

1. Extrae's build system had a couple of bugs in order to make it work with BinaryBuilder
2. It was the first Yggdrasil recipe that had **both** CUDA and MPI platform augmentations

> No installation instructions. Not for Extrae.jl and not for the parent tool (at least a link?).
> "Example usage": Would be great to have a few more examples (e.g. multithreading, MPI).
> Expand statement of need: Why do we need Extrae, specifically? Why aren't the other, mentioned tools not sufficient?
> Link to JuliaCon24 talk in the README.md?

I'm leaving here a list of tasks to do:

- [ ] Add examples
  - [ ] multithread
  - [ ] MPI
- [ ] Update README.md with JuliaCon24 talk and installation instructions
- [ ] Add statement of need
- [ ] Apply proposed changes and corrections
