### 2025-04-16 c0a86ea
Add processed SVG plot
