### 2025-05-27 4fb04ac
Add "Known issues" section
