### 2025-05-27 #180 @luraess
> statement of need. is this a section i need to add to the article?

Thanks for bringing this up. No, you do not need to explicitly include this section in the article. However, this information should be implicitly included in the motivation of your work. We will adapt the check list soon to reflect this.
