### 2025-05-27 #180 @mofeing
Examples added in the repo. I believe that the only thing left is the statement of need.

From @carstenbauer's review, I guess the article needs to emphasize on this?

> Expand statement of need: Why do we need Extrae, specifically? Why aren't the other, mentioned tools not sufficient?

I'm a lil bit confused on how to answer this. Extrae is not a new development, and there are other similar packages in the Julia ecosystem like LIKWID.jl or LinuxPerf.jl (in the sense that they are binding packages to existing HPC profiling software). The features that differentiate, for example, Extrae.jl and LIKWID.jl, are the features that differentiate the original Extrae and LIKDWID profilers.
