### 2025-06-24 #180 @JBlaschke
Ok I've gone over the manuscript -- I think it's very easy to follow (in a good way), and I like it a lot. There are several issues that I would like @mofeing and/or @giordano to address:

- [ ] I have labeled where more citations are needed
- [ ] This is an HPC profiler, and @mofeing and @giordano have already done a great job at starting the discussion about the whole `MPI.jl` preload shenanigans we need to do if you look here: https://github.com/bsc-quantic/Extrae.jl/blob/master/examples/mpi/LocalPreferences.toml -- HOWEVER I suspect that I found this only because I am looking for it. So the documentation (and to some extend the paper) needs to unambiguously explain: i) when you can use `Extrae_jll` (and does that work with `MPItrampoline`?); ii) when a system install of `Extrae` is needed (e.g. if using CrayMPICH?);
- [ ] I suggest provided a more detailed example of how to use each of these cases -- i.e. how to use the jll, and how to use a system MPI, and therefore a pre-compiled version of `Extae`
- [ ] The instructions [here](https://tools.bsc.es/doc/pdf/extrae-installation-guide.pdf) for how to build `Extrae` yourself are a good start, but again I suspect this is because I'm familiar with HPC (and even then it took half a day). I added an automatic build to the way I manage the [NERSC Julia Install](https://github.com/JuliaParallel/DeploymentsOnHPC) (`nersc/extrae/sm-config/pre_install.sh` installs all the dependencies; and `nersc/extrae/sm-config/install.sh` installs extrae itself) -- I suggest using something like that to create an example `install-everything.sh` script.

In truly old-school fashion, I am attaching a scanned doc with hand-written notes: [extrae_jl_review.pdf](https://github.com/user-attachments/files/20875847/extrae_jl_review.pdf)
