### 2025-11-12 #180 @luraess
> thank you for the review [@JBlaschke](https://github.com/JBlaschke). please, contact me if you have some issues when trying it on perlmutter.
>
> [@luraess](https://github.com/luraess) will add the corrections as soon as I find some free time, which should be in a matter of weeks

Can we get this moving forward, or if you feel this submission is no longer a thing, let me know so we can terminate the process. It's now soon half a year without any action, reply nor final fixes. I'll set a reminder and if no answers nor actions are undertaken within the next 2 weeks, I'll consider this submission as stall and will see to close it @mofeing
